@@ -34,7 +34,7 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.3",
         "ext-filter": "*",
         "ext-hash": "*",
         "ext-json": "*",
@@ -44,14 +44,14 @@
         "alibabacloud/cdn": "^1.7",
         "alibabacloud/wafopenapi": "^1.7",
         "aws/aws-sdk-php": "^3.38",
+        "consolidation/self-update": "^2.2",
         "guzzlehttp/guzzle": "^7.2",
-        "guzzlehttp/psr7": "^1.0",
-        "lcobucci/jwt": "^3.3|^4.0",
+        "guzzlehttp/psr7": "^2.4.5",
+        "lcobucci/jwt": "^5.3",
         "league/flysystem": "^3.10",
         "league/flysystem-memory": "^3.10",
         "league/flysystem-sftp-v3": "^3.10",
         "monolog/monolog": "^1.19|^2.0",
-        "padraic/phar-updater": "^1.0",
         "psr/container": "^1.0",
         "psr/http-message": "^1.0",
         "psr/log": "^1.0",
@@ -61,10 +61,10 @@
         "symfony/filesystem": "^5.4.12 || ^6.4",
         "symfony/serializer": "^5.4.12 || ^6.4",
         "symfony/yaml": "^5.4.12 || ^6.4",
-        "webmozart/assert": "^1.0",
-        "webmozart/path-util": "^2.3"
+        "webmozart/assert": "^1.0"
     },
     "require-dev": {
+        "phpstan/phpstan": "^1.11",
         "symfony/finder": "^5.4.12 || ^6.4",
         "symfony/phpunit-bridge": "^5.4.12 || ^6.4",
         "symfony/property-access": "^5.4.12 || ^6.4",
