@@ -14,11 +14,11 @@ namespace AcmePhp\Cli;
 use AcmePhp\Cli\Command\Helper\DistinguishedNameHelper;
 use AcmePhp\Cli\Command\RevokeCommand;
 use AcmePhp\Cli\Command\RunCommand;
-use AcmePhp\Cli\Command\SelfUpdateCommand;
 use AcmePhp\Cli\Command\StatusCommand;
+use SelfUpdate\SelfUpdateCommand;
 use Symfony\Component\Console\Application as BaseApplication;
 use Symfony\Component\Console\Helper\HelperSet;
-use Webmozart\PathUtil\Path;
+use Symfony\Component\Filesystem\Path;
 
 /**
  * @author Titouan Galopin <galopintitouan@gmail.com>
@@ -33,17 +33,20 @@ class Application extends BaseApplication
 
     public function __construct()
     {
-        parent::__construct('Acme PHP - Let\'s Encrypt/ZeroSSL client', '3.0.0');
+        // This is replaced by humbug/box with a string that looks like this: x.y.z@tag
+        parent::__construct('Acme PHP - Let\'s Encrypt/ZeroSSL client', '@git@');
     }
 
     protected function getDefaultCommands(): array
     {
-        return array_merge(parent::getDefaultCommands(), array(
+        $version = explode('@', $this->getVersion())[0];
+
+        return array_merge(parent::getDefaultCommands(), [
             new RunCommand(),
             new RevokeCommand(),
             new StatusCommand(),
-            new SelfUpdateCommand(),
-        ));
+            new SelfUpdateCommand($this->getName(), '' === $version ? '0.0.0' : $version, 'acmephp/acmephp'),
+        ]);
     }
 
     protected function getDefaultHelperSet(): HelperSet
